Move GitRepositorySandbox to its own file
