+++ tests/test_advance.py
@@ -3,10 +3,11 @@ from pytest_mock import MockerFixture
 
 from git_machete.exceptions import UnderlyingGitException
 
-from .base_test import BaseTest, GitRepositorySandbox
+from .base_test import BaseTest
 from .mockers import (assert_failure, assert_success, launch_command,
                       mock_input_returning, mock_input_returning_y,
                       rewrite_branch_layout_file)
+from .mockers_git_repo_sandbox import GitRepositorySandbox
 
 
 class TestAdvance(BaseTest):
