+++ tests/test_anno.py
@@ -1,9 +1,10 @@
 from pytest_mock import MockerFixture
 
 from . import mockers_github, mockers_gitlab
-from .base_test import BaseTest, GitRepositorySandbox
+from .base_test import BaseTest
 from .mockers import (assert_failure, assert_success, launch_command,
                       rewrite_branch_layout_file)
+from .mockers_git_repo_sandbox import GitRepositorySandbox
 from .mockers_github import (MockGitHubAPIState,
                              mock_github_token_for_domain_fake)
 from .mockers_gitlab import (MockGitLabAPIState,
