+++ tests/test_diff.py
@@ -1,7 +1,8 @@
 from pytest_mock import MockerFixture
 
-from .base_test import BaseTest, GitRepositorySandbox
+from .base_test import BaseTest
 from .mockers import assert_success, mock__run_cmd_and_forward_stdout
+from .mockers_git_repo_sandbox import GitRepositorySandbox
 
 
 class TestDiff(BaseTest):
