+++ tests/test_file.py
@@ -4,8 +4,9 @@ from tempfile import mkdtemp
 
 from git_machete.exceptions import UnderlyingGitException
 
-from .base_test import BaseTest, GitRepositorySandbox
+from .base_test import BaseTest
 from .mockers import assert_failure, launch_command
+from .mockers_git_repo_sandbox import GitRepositorySandbox
 
 
 class TestFile(BaseTest):
