+++ tests/test_gitlab_restack_mr.py
@@ -2,11 +2,12 @@ import textwrap
 
 from pytest_mock import MockerFixture
 
-from tests.base_test import BaseTest, GitRepositorySandbox
+from tests.base_test import BaseTest
 from tests.mockers import (assert_failure, assert_success,
                            fixed_author_and_committer_date_in_past,
                            rewrite_branch_layout_file)
 from tests.mockers_code_hosting import mock_from_url
+from tests.mockers_git_repo_sandbox import GitRepositorySandbox
 from tests.mockers_gitlab import (MockGitLabAPIState,
                                   mock_gitlab_token_for_domain_fake,
                                   mock_mr_json, mock_urlopen)
