+++ tests/test_go.py
@@ -3,9 +3,10 @@ from pytest_mock import MockerFixture
 
 from git_machete.exceptions import ExitCode
 
-from .base_test import BaseTest, GitRepositorySandbox
+from .base_test import BaseTest
 from .mockers import (assert_failure, assert_success, launch_command,
                       mock_input_returning, rewrite_branch_layout_file)
+from .mockers_git_repo_sandbox import GitRepositorySandbox
 
 
 class TestGo(BaseTest):
