+++ tests/test_help.py
@@ -5,8 +5,9 @@ import pytest
 from git_machete.cli import commands_and_aliases
 from git_machete.exceptions import ExitCode
 
-from .base_test import BaseTest, GitRepositorySandbox
+from .base_test import BaseTest
 from .mockers import launch_command
+from .mockers_git_repo_sandbox import GitRepositorySandbox
 
 
 class TestHelp(BaseTest):
