+++ tests/test_list.py
@@ -2,9 +2,10 @@ import pytest
 
 from git_machete.exceptions import ExitCode
 
-from .base_test import BaseTest, GitRepositorySandbox
+from .base_test import BaseTest
 from .mockers import (assert_failure, assert_success, launch_command,
                       rewrite_branch_layout_file)
+from .mockers_git_repo_sandbox import GitRepositorySandbox
 
 
 class TestList(BaseTest):
