+++ tests/test_slide_out.py
@@ -1,11 +1,12 @@
 import pytest
 from pytest_mock import MockerFixture
 
-from .base_test import BaseTest, GitRepositorySandbox
+from .base_test import BaseTest
 from .mockers import (assert_failure, assert_success,
                       fixed_author_and_committer_date_in_past, launch_command,
                       mock_input_returning_y, read_branch_layout_file,
                       rewrite_branch_layout_file)
+from .mockers_git_repo_sandbox import GitRepositorySandbox
 
 
 class TestSlideOut(BaseTest):
