+++ tests/test_status.py
@@ -5,11 +5,12 @@ import textwrap
 import pytest
 from pytest_mock import MockerFixture
 
-from .base_test import BaseTest, GitRepositorySandbox
+from .base_test import BaseTest
 from .mockers import (assert_failure, assert_success,
                       fixed_author_and_committer_date_in_past, launch_command,
                       mock_input_returning, mock_input_returning_y,
                       overridden_environment, rewrite_branch_layout_file)
+from .mockers_git_repo_sandbox import GitRepositorySandbox
 
 
 class TestStatus(BaseTest):
@@ -877,7 +878,8 @@ class TestStatus(BaseTest):
 
     def test_status_removed_from_remote(self) -> None:
         (
-            self.repo_sandbox.new_branch('main')
+            GitRepositorySandbox()
+            .new_branch('main')
             .commit()
             .push()
             .delete_remote_branch('origin/main')
